% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{fetch_all}
\alias{fetch_all}
\title{Fetch all of the data}
\usage{
fetch_all(sites = read_sites()$name, dst_path = ".")
}
\arguments{
\item{sites}{vector of site names (short names)}

\item{dst_path}{the destuntion path to save the files}
}
\value{
vector of 0 = success, !0 = failure
}
\description{
Fetch all of the data
}
