% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout-package.R
\name{iceout}
\alias{iceout}
\alias{umoos}
\title{iceout-package}
\description{
Provides convenient access to New England iceout data from \url{https://me.water.usgs.gov/iceout_data}.
}
\seealso{
\url{https://me.water.usgs.gov/iceout_data}
}

