% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{iceout_path}
\alias{iceout_path}
\title{Retrieve the data path for the package data}
\usage{
iceout_path(...)
}
\arguments{
\item{...}{path elements}
}
\value{
charcater path
}
\description{
Retrieve the data path for the package data
}
