% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\docType{data}
\name{iceout_uri}
\alias{iceout_uri}
\title{Retrieve the URI for online data access}
\format{An object of class \code{character} of length 1.}
\usage{
iceout_uri
}
\value{
charcater uri
}
\description{
Retrieve the URI for online data access
}
\keyword{internal}
