% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mn_iceout}
\alias{mn_iceout}
\title{Minnesota Department of Natural Resources State Ice-out Data}
\format{A data frame with 19,261 rows and 9 variables}
\usage{
data(mn_iceout)
}
\description{
This dataset contains the Minnesota state-level historical data
generated via \code{\link[=read_mn_iceout_data]{read_mn_iceout_data()}}.
\itemize{
\item \code{source} <chr> (recording person/institution)
\item \code{name} <chr> (inland body of water name)
\item \code{date} <Date> (ice-out date)
\item \code{comments} <chr> (field notes)
\item \code{lat}/\code{lon} <dbl> (coordinates)
\item \code{sentinel_lake} <int> (see site)
\item \code{state} <chr> (always Minnesota)
}
}
\note{
Last updated 2019-01-15.
}
\references{
\url{https://www.dnr.state.mn.us/ice_out/index.html?year=1843}
}
\keyword{datasets}
