% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nsidc_iceout}
\alias{nsidc_iceout}
\title{NSIDC Global Lake and River Ice Phenology Database, Version 1}
\format{A data frame with 35,918 rows and 37 variables}
\usage{
data(nsidc_iceout)
}
\description{
The Global Lake and River Ice Phenology Database contains
freeze and thaw/breakup dates as well as other descriptive ice cover data
for 865 lakes and rivers in the Northern Hemisphere. Of the 542 water
bodies that have records longer than 19 years, 370 of them are in North
America and 172 are in Eurasia. 249 lakes and rivers have records longer
than 50 years, and 66 have records longer than 100 years. A few water
bodies have data available prior to 1845. This database, with water
bodies distributed around the Northern Hemisphere, allows for the
analysis of broad spatial patterns as well as long-term temporal patterns.
\itemize{
\item \code{lakecode} <chr> ()
\item \code{lakename} <chr> ()
\item \code{lakeorriver} <chr> ()
\item \code{season} <chr> ()
\item \code{iceon_year} <dbl> ()
\item \code{iceon_month} <dbl> ()
\item \code{iceon_day} <dbl> ()
\item \code{iceoff_year} <dbl> ()
\item \code{iceoff_month} <dbl> ()
\item \code{iceoff_day} <dbl> ()
\item \code{duration} <dbl> ()
\item \code{latitude} <lgl> ()
\item \code{longitude} <lgl> ()
\item \code{country} <chr> ()
\item \code{froze} <lgl> ()
\item \code{obs_comments} <chr> ()
\item \code{area_drained} <lgl> ()
\item \code{bow_comments} <chr> ()
\item \code{conductivity_us} <dbl> ()
\item \code{elevation} <lgl> ()
\item \code{filename} <chr> ()
\item \code{initials} <chr> ()
\item \code{inlet_streams} <chr> ()
\item \code{landuse_code} <chr> ()
\item \code{largest_city_population} <lgl> ()
\item \code{max_depth} <lgl> ()
\item \code{mean_depth} <lgl> ()
\item \code{median_depth} <lgl> ()
\item \code{power_plant_discharge} <lgl> ()
\item \code{secchi_depth} <lgl> ()
\item \code{shoreline} <lgl> ()
\item \code{surface_area} <lgl> ()
\item \code{state} <chr> ()
\item \code{iceon_date} <date ()
\item \code{iceon_doy} <dbl> ()
\item \code{iceout_date} <date ()
\item \code{iceout_doy} <dbl> ()
}
}
\note{
Last updated 2019-01-18.
}
\references{
\url{http://nsidc.org/data/lake_river_ice/}
}
\keyword{datasets}
