% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{parse_all}
\alias{parse_all}
\title{Parse one or more sites}
\usage{
parse_all(sites = read_sites()$name)
}
\arguments{
\item{sites}{one or more site names (short name)}
}
\value{
list of iceout lists
}
\description{
Parse one or more sites
}
\seealso{
Other dataset builders: \code{\link{fetch_all}},
  \code{\link{iceout_path}}, \code{\link{parse_iceout}},
  \code{\link{read_iceout}},
  \code{\link{read_maine_iceout_data}},
  \code{\link{read_mn_iceout_data}},
  \code{\link{read_sites}}
}
\concept{dataset builders}
\keyword{internal}
