% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{parse_iceout}
\alias{parse_iceout}
\title{Parse a single iceout file}
\usage{
parse_iceout(name = "Auburn", form = "tibble")
}
\arguments{
\item{name}{the site name}

\item{form}{character, just 'tibble' for now but possibly 'sf' in future}
}
\value{
list of iceout metadata and data in \code{form} form
}
\description{
Parse a single iceout file
}
\seealso{
Other dataset builders: \code{\link{fetch_all}},
  \code{\link{iceout_path}}, \code{\link{parse_all}},
  \code{\link{read_iceout}},
  \code{\link{read_maine_iceout_data}},
  \code{\link{read_mn_iceout_data}},
  \code{\link{read_sites}}
}
\concept{dataset builders}
