% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_iceout}
\alias{plot_iceout}
\title{Plot a single iceout dataset (excluding data from \code{nsidc_icout})}
\usage{
plot_iceout(dataset = c("usgs", "me", "mn"), state = "Maine",
  body_name = "Auburn", town = NULL, method = "loess", formula = y
  ~ x, ...)
}
\arguments{
\item{dataset}{one of "\code{usgs}" for the core USGS data, "\code{me}" for the Maine
state curated data, or "\code{mn}" for the Minnesota state curated data.}

\item{state}{a valid state name for \code{dataset}}

\item{body_name}{a valid inland body of water name for the \code{dataset} and
\code{state} combination.}

\item{town}{the Maine state curated data includes ponds and pond names are
not unique across the state so if specifying "\code{me}" for \code{dataset}
this parameter must be specified.}

\item{method, formula, ...}{see \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}}
}
\value{
ggplot2 object
}
\description{
Plot a single iceout dataset (excluding data from \code{nsidc_icout})
}
\note{
This function will not pull data from the \code{nsidc_icout} dataset.
}
\examples{
plot_iceout("usgs", "New Hampshire", "First.Conn")
plot_iceout("me", "Maine", "Long Pond", town = "Belgrade")
plot_iceout("mn", "Minnesota", "Minnewaska")
}
