% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{read_iceout}
\alias{read_iceout}
\title{Read the prepared iceout data}
\usage{
read_iceout(filename = iceout_path("iceout_data.rds"))
}
\arguments{
\item{filename}{the name of the iceout_data.rds file}
}
\value{
a list of parsed iceout datasets
}
\description{
Read the prepared iceout data
}

