% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout-me.R
\name{read_maine_iceout_data}
\alias{read_maine_iceout_data}
\title{Read Maine Bureau of Parks & Lands State Ice-out Data for a Given Year}
\usage{
read_maine_iceout_data(year = 2003)
}
\arguments{
\item{year}{= 2003 (defaults to 2003)}
}
\value{
data frame (tibble) consisting of \code{state}, \code{town}, \code{body_name}, \code{date},
\code{year}, \code{doy}
}
\description{
The state of Maine maintains its own data (going back to 2003) of
ice-out data for inland lakes and ponds. Since the \href{https://me.water.usgs.gov/iceout_data}{USGS data}
only goes up to 2005, we can use the Maine state data to retrieve  measurements
up to present year of available ice-out data.
}
\details{
This function will pull data directly from the Maine Bureau of Parks & Lands
web pages (e.g. \url{https://www.maine.gov/dacf/parks/water_activities/boating/ice_out06.shtml}).
}
\examples{
me_iceout <- read_maine_iceout_data(2003)
}
\seealso{
Other online readers: \code{\link{read_mdnr_metadata}},
  \code{\link{read_minnesota_iceout_data}}
}
\concept{online readers}
