% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout-mn.R
\name{read_mdnr_metadata}
\alias{read_mdnr_metadata}
\title{Retrieves metadata on all available Minnesota Lakes.}
\usage{
read_mdnr_metadata()
}
\value{
data frame
}
\description{
Use this to get the \code{id} for \code{\link[=read_minnesota_iceout_data]{read_minnesota_iceout_data()}}
}
\examples{
meta <- read_mdnr_metadata()
meta_u <- meta[, c("name", "lat", "lon", "id")]
meta_u <- meta_u[!duplicated(meta_u),]
meta_u <- meta_u[order(meta_u$name),]

\dontrun{
# tidyverse
library(dplyr)
select(meta, name, lat, lon, id) \%>\%
  arrange(name)
}
}
\seealso{
Other online readers: \code{\link{read_maine_iceout_data}},
  \code{\link{read_minnesota_iceout_data}}
}
\concept{online readers}
