% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout-mn.R
\name{read_minnesota_iceout_data}
\alias{read_minnesota_iceout_data}
\title{Read Minnesota Department of Natural Resources State Ice-out Data}
\usage{
read_minnesota_iceout_data(id)
}
\arguments{
\item{id}{\code{<chr>} the MDNR identifier of the body of water. See \code{\link[=read_mdnr_metadata]{read_mdnr_metadata()}}.}
}
\value{
data frame
}
\description{
The state of Minnesota maintains its own data (going back to 1843 and up to present-day) of
ice-out data for inland lakes and ponds. Since the \href{https://me.water.usgs.gov/iceout_data}{USGS data}
only covers New England ice-outs, we can use this site to gather data
for Minnesota to supplement analyses.
}
\details{
This function will pull data directly from the Minnesota Department of Natural Resources
web pages (e.g. \url{https://www.dnr.state.mn.us/ice_out/index.html?year=1843}) to
find generate the data links to retrieve the data.
}
\note{
You get \emph{all} historical records for a given MDNR id. Identifiers can
begin with a \code{0} so make sure this is a character.
}
\examples{
read_minnesota_iceout_data("11030500")
}
\seealso{
Other online readers: \code{\link{read_maine_iceout_data}},
  \code{\link{read_mdnr_metadata}}
}
\concept{online readers}
