% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout-mn.R
\name{read_mn_iceout_data}
\alias{read_mn_iceout_data}
\title{Read Minnesota Department of Natural Resources State Ice-out Data}
\usage{
read_mn_iceout_data()
}
\value{
data frame (tibble) consisting of \code{source}, \code{body_name} and \code{iceout_date}
}
\description{
The state of Minnesota maintains its own data (going back to 1843 and up to present-day) of
ice-out data for inland lakes and ponds. Since the \href{https://me.water.usgs.gov/iceout_data}{USGS data}
only covers New England ice-outs, we can use this site to gather data
for Minnesota to supplement analyses.
}
\details{
This function will pull data directly from the Minnesota Department of Natural Resources
web pages (e.g. \url{https://www.dnr.state.mn.us/ice_out/index.html?year=1843}) to
find generate the data links to retrieve the data.
}
\note{
this is really for bootstrapping the built-in data set. You are
better off doing \code{data(mn_iceout)} vs hitting up the MDNR servers.
}
\examples{
\dontrun{
mn_iceout <- read_mn_iceout_data()
}
}
\seealso{
Other dataset builders: \code{\link{fetch_all}},
  \code{\link{iceout_path}}, \code{\link{parse_all}},
  \code{\link{parse_iceout}}, \code{\link{read_iceout}},
  \code{\link{read_maine_iceout_data}},
  \code{\link{read_sites}}
}
\concept{dataset builders}
\keyword{internal}
