% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{read_sites}
\alias{read_sites}
\title{Read the iceout_sites file that provides name, longname and location data}
\usage{
read_sites(filename = iceout_path("iceout_sites.csv"))
}
\arguments{
\item{filename}{the fully qualified filename}
}
\value{
tibble of site data
}
\description{
Read the iceout_sites file that provides name, longname and location data
}

