% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{usgs_iceout}
\alias{usgs_iceout}
\title{USGS Lake Ice-Out Data for New England}
\format{A data frame with 3,372 rows and 9 variables}
\usage{
data(usgs_iceout)
}
\description{
This dataset contains the USGS historical ice-out data
for New England
\itemize{
\item \code{state} <chr> (state lake is in)
\item \code{name} <chr> (short lake namne)
\item \code{long_name} <chr> (long lake name)
\item \code{lat}/\code{lon} <dbl> (coordinates)
\item \code{year} <int> (ice-out year)
\item \code{doy} <int> (ice-out day of year)
\item \code{date} <Date> (ice-out date)
\item \code{observer} <chr> (recording person/institution)
}
}
\note{
Umbagog lake spans New Hampshire & Maine and is in the data
frame under both states.

Last updated 2019-01-15.
}
\references{
\url{https://me.water.usgs.gov/iceout_data/};
\url{https://me.water.usgs.gov/iceout.html}
}
\keyword{datasets}
