% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{iceout_path}
\alias{iceout_path}
\title{Retrieve the data path for the package data}
\usage{
iceout_path(...)
}
\arguments{
\item{...}{path elements}
}
\value{
charcater path
}
\description{
Retrieve the data path for the package data
}
\seealso{
Other dataset builders: \code{\link{fetch_all}},
  \code{\link{parse_all}}, \code{\link{parse_iceout}},
  \code{\link{read_iceout}},
  \code{\link{read_maine_iceout_data}},
  \code{\link{read_mn_iceout_data}},
  \code{\link{read_sites}}
}
\concept{dataset builders}
\keyword{internal}
