% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{parse_all}
\alias{parse_all}
\title{Parse one or more sites}
\usage{
parse_all(sites = read_sites()$name)
}
\arguments{
\item{sites}{one or more site names (short name)}
}
\value{
list of iceout lists
}
\description{
Parse one or more sites
}
