% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{plot_iceout}
\alias{plot_iceout}
\title{Plot a single iceout dataset}
\usage{
plot_iceout(x = read_iceout()[[1]], trend = "loess")
}
\arguments{
\item{x}{a list of iceout data}

\item{trend}{charcater, if 'loess' then draw a loess fit on the data;
other values are ignored for now.}
}
\value{
ggplot2 object
}
\description{
Plot a single iceout dataset
}
