% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout.R
\name{read_iceout}
\alias{read_iceout}
\title{Read the prepared iceout data}
\usage{
read_iceout(filename = iceout_path("iceout_data.rds"))
}
\arguments{
\item{filename}{the name of the iceout_data.rds file}
}
\value{
a list of parsed iceout datasets
}
\description{
Read the prepared iceout data
}
\seealso{
Other dataset builders: \code{\link{fetch_all}},
  \code{\link{iceout_path}}, \code{\link{parse_all}},
  \code{\link{parse_iceout}},
  \code{\link{read_maine_iceout_data}},
  \code{\link{read_mn_iceout_data}},
  \code{\link{read_sites}}
}
\concept{dataset builders}
\keyword{internal}
