% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceout-maine.R
\name{read_maine_iceout_data}
\alias{read_maine_iceout_data}
\title{Read Maine Bureau of Parks & Lands State Ice-out Data}
\usage{
read_maine_iceout_data()
}
\value{
data frame (tibble) consisting of \code{town}, \code{body_name} and \code{iceout_date}
}
\description{
The state of Maine maintains its own data (going back to 2003) of
ice-out data for inland lakes and ponds. Since the \href{https://me.water.usgs.gov/iceout_data}{USGS data}
only goes up to 2005, we can use this data to extend the measurement period
to present year.
}
\details{
This function will pull data directly from the Maine Bureau of Parks & Lands
web pages (e.g. \url{https://www.maine.gov/dacf/parks/water_activities/boating/ice_out06.shtml})
and there is a current hardcoded "end year" of 2017.
}
\note{
this takes a while and you are better off using \code{data(me_icout)} since
it will be updated each season. Only use this if you're impatient.
}
\examples{
\dontrun{
me_iceout <- read_maine_iceout_data()
}
}
